Okay, so I just had to fking deal with debugging a plugin which is not a very fun way to waste my ever so scarce free time. THEN! God itself had to butt fk me with this blasphemous behavioural shit with this fjhwebfhujebfghkr fvj code! please wai....
I tried adding a forward/skip function for the cursor, in which it would
forward the cursor and render the data on screen immediately for the
next node, but noooooooooooooooooooooooooooo. It just had to that the
Engine decides it would emit my signal forward twice from who knows where,
forwarding all my rendered data until the fking stone age. I tried to debug it all day,
but i cant find a single damn reason for that signal to be called twice!
when i am just calling it once.... And I can the fuckk not stand using Godot's
command line debugger any further. It is a torturous, and bad for the soul.

In what universe does a command line debugger not have the feature to
list my current file so i can see where the fuck i am currently
at!!!#12iewnfjdwfdf I AM DEAD.