Cursed lifted!
Okay I found out the source of the curse... Turns out, the reason why
every exported resource variable is shared on each instances is because
when their _ready() function is triggered, which some may assigned those
exported variables into a new resource. Then when I close Godot, those
scene would use those resources instead of the default assignment i did
with:
  export var name : Resource = null.

So it is not necessary to default assign those resources to null.